changed distance type to double, added to Runner tests

+++ RunDiary/Models/Run.cs
@@ -17,9 +17,9 @@ namespace RunDiary.Models
         public string RunName { get; set; }
         public string RunPlace { get; set; }
         public DateTime RunDate { get; set; }
-        public int RunDistance { get; set; }
+        public double RunDistance { get; set; }
         public string DistanceUnits { get; set; }
-        public int RunTime { get; set; }
+        public double RunTime { get; set; }
         public string RunTimeUnits { get; set; }
         public string Photo { get; set; }
         public string DiaryEntry { get; set; }
